engine: extract esm/wwt_control.js
